# Improve the efficiency of silence