This makes `TakeDuration` cache the duration per sample calculation and adjusts the silence sample rate to what appears to be the default so that the sample rate doesn't need to be adjusted.