@Imberflur 2020-02-23
I can get some criterion benchmark data.