@Imberflur 2020-02-24
Before the change
```
dynamic mixer with 1 silent source                                                                             
                        time:   [22.000 ns 22.415 ns 22.982 ns]

dynamic mixer with 10 silent sources                                                                            
                        time:   [238.63 ns 257.99 ns 276.70 ns]
```
With the changes to `TakeDuration`
```
dynamic mixer with 1 silent source                                                                             
                        time:   [19.129 ns 19.946 ns 21.003 ns]
                        change: [-22.109% -15.749% -8.5357%] (p = 0.00 < 0.05)
                        Performance has improved.

dynamic mixer with 10 silent sources                                                                            
                        time:   [191.22 ns 207.62 ns 221.65 ns]
                        change: [-29.447% -23.587% -17.037%] (p = 0.00 < 0.05)
                        Performance has improved.
```
Building on that by adjusting the silence sample rate to be that of the mixer (44100 hz)
```
dynamic mixer with 1 silent source                                                                             
                        time:   [14.955 ns 16.052 ns 17.339 ns]
                        change: [-40.851% -35.805% -30.248%] (p = 0.00 < 0.05)
                        Performance has improved.

dynamic mixer with 10 silent sources                                                                            
                        time:   [104.07 ns 114.65 ns 124.49 ns]
                        change: [-46.665% -42.486% -37.811%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Bench function
```rust
pub fn criterion_benchmark(c: &mut Criterion) {
    let (mixer_tx, mut mixer_rx) = dynamic_mixer::mixer::<f32>(2, 44100);
    let _ = {  
        let (queue_tx, queue_rx) = queue::queue(true);
        mixer_tx.add(queue_rx);
        queue_tx 
    };
    c.bench_function("dynamic mixer with 1 silent source", |b| b.iter(|| mixer_rx.next()));
    
    let (mixer_tx, mut mixer_rx) = dynamic_mixer::mixer::<f32>(2, 44100);
    let _ = (0..).take(10).map(|_| {  
        let (queue_tx, queue_rx) = queue::queue(true);
        mixer_tx.add(queue_rx);
        queue_tx 
    }).collect::<Vec<_>>();
    c.bench_function("dynamic mixer with 10 silent sources", |b| b.iter(|| mixer_rx.next()));
}
```