@est31 2020-02-24
Thanks for doing the benchmarks. It's surprising that these changes have so large impacts. I'd like to merge the PR now. For that, it needs to be rebased. Could you please rebase it?